pyproject.toml
[tool.poetry]
name = "cartiflette"
version = "0.0.2"
description = "Tools to easily retrieve French borders for geographic data analytics"
authors = ["Lino Galiana <lino.galiana@insee.fr>", "Thomas Grandjean <thomas.grandjean@developpement-durable.gouv.fr>"]
license = "OPEN LICENCE 2.0/LICENCE OUVERTE 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
py7zr = "^0.20.4"
geopandas = "^0.12.2"
openpyxl = "^3.1.1"
tqdm = "^4.65.0"
s3fs = "^2023.3.0"
PyYAML = "^6.0"
xlrd = "^2.0.1"
topojson = "^1.5"
appdirs = "^1.4.4"
owslib = "^0.28.1"
matplotlib = "^3.7.1"
python-dotenv = "^1.0.0"
python-magic-bin = {version = "^0.4.14", markers = "sys_platform == 'win32'"}
python-magic = {version = "^0.4.14", markers = "sys_platform == 'linux'"}
requests-cache = "^1.1.0"
unidecode = "^1.3.6"
charset-normalizer = "^3.2.0"
pebble = "^5.0.3"
fiona = {extras = ["s3"], version = "^1.9.4.post1"}

[tool.poetry.group.dev.dependencies]
spyder = "^5.4.2"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"